Refactor GeoTIFF Phase 5d: extract _cog_http.py from _reader.py (#2258)
Move the COG-over-HTTP transport layer (`_parse_cog_http_meta`,
`_read_cog_http`, `_fetch_decode_cog_http_strips`,
`_fetch_decode_cog_http_tiles`) plus the `INITIAL_HTTP_HEADER_BYTES` /
`MAX_HTTP_HEADER_BYTES` constants into a new `_cog_http.py`. `_reader.py`
drops from 1258 to 429 lines.

The public import surface stays intact: `_reader.py` re-exports the
moved names so `from xrspatial.geotiff._reader import _read_cog_http`
and friends keep working unchanged.

`_read_cog_http` resolves `_HTTPSource`, `_parse_cog_http_meta`,
`_fetch_decode_cog_http_tiles`, `_decode_strip_or_tile`, and
`_apply_photometric_miniswhite` through the `_reader` module namespace
at call time so existing tests that
`monkeypatch.setattr(_reader, '_HTTPSource', ...)` (and friends)
continue to intercept the collaborators. `_parse_cog_http_meta` reads
`MAX_HTTP_HEADER_BYTES` and `INITIAL_HTTP_HEADER_BYTES` the same way so
`test_http_meta_buffer_1718`'s cap-shrinking patch keeps working.

Part of #2211.
